Split DragAndScale out to TS file
